17.1. all done

+++ Libft/ft_strlcat.c
@@ -6,16 +6,16 @@
 /*   By: thbui <thbui@student.42prague.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 22:39:41 by thbui             #+#    #+#             */
-/*   Updated: 2023/01/15 12:46:30 by thbui            ###   ########.fr       */
+/*   Updated: 2023/01/17 19:17:21 by thbui            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = ft_strlen(dest);
