+++ Libft/ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: thbui <thbui@student.42prague.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 20:38:27 by thbui             #+#    #+#             */
-/*   Updated: 2023/01/15 12:46:37 by thbui            ###   ########.fr       */
+/*   Updated: 2023/01/17 19:18:02 by thbui            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (size != 0)
